@@ -2,17 +2,20 @@ import java.net.URI
 
 plugins {
 	id("maven-publish")
-	id("fabric-loom") version "1.11.7"
-	id("babric-loom-extension") version "1.10.2"
+	id("net.fabricmc.fabric-loom-remap") version "1.15.+"
+	id("ploceus") version "1.15-SNAPSHOT"
 }
 
-//noinspection GroovyUnusedAssignment
-java.sourceCompatibility = JavaVersion.VERSION_17
-java.targetCompatibility = JavaVersion.VERSION_17
+java.sourceCompatibility = JavaVersion.VERSION_21
+java.targetCompatibility = JavaVersion.VERSION_21
 
 base.archivesName = project.properties["archives_base_name"] as String
 version = project.properties["mod_version"] as String
 group = project.properties["maven_group"] as String
+
+ploceus {
+	setIntermediaryGeneration(2)
+}
 
 loom {
 	accessWidenerPath = file("src/main/resources/gcapi3.accesswidener")
