@@ -4,9 +4,9 @@ import net.glasslauncher.mods.gcapi3.impl.GCCore;
 import net.glasslauncher.mods.gcapi3.impl.object.ConfigCategoryHandler;
 import net.glasslauncher.mods.gcapi3.impl.object.ConfigEntryHandler;
 import net.glasslauncher.mods.networking.GlassNetworking;
-import net.minecraft.class_454;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.world.ClientWorld;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -19,9 +19,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class MinecraftMixin {
     @Shadow public World world;
 
-    @Inject(method = "method_2115", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/player/ClientPlayerEntity;method_1315()V", ordinal = 1))
+    @Inject(method = "setWorld(Lnet/minecraft/world/World;Ljava/lang/String;Lnet/minecraft/entity/player/PlayerEntity;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/player/ClientPlayerEntity;teleportTop()V"))
     private void checkVanillaJoined(World string, String playerEntity, PlayerEntity par3, CallbackInfo ci) {
-        if(!GlassNetworking.serverHasNetworking() && world instanceof class_454) {
+        if(!GlassNetworking.serverHasNetworking() && world instanceof ClientWorld) {
             //noinspection deprecation
             GCCore.MOD_CONFIGS.forEach((identifier, configRootEntry) -> recursiveTriggerVanillaBehavior(configRootEntry.configCategoryHandler()));
         }
