@@ -9,6 +9,7 @@ import org.jetbrains.annotations.ApiStatus;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.function.BiConsumer;
@@ -73,15 +74,15 @@ public class GlassPacket extends Packet {
     }
 
     @Override
-    public void write(DataOutputStream stream) {
+    public void write(DataOutputStream stream) throws IOException {
         nbt.putString("glassnetworking:packetId", packetId);
         nbt.putString("glassnetworking:modId", modId);
-        length = net.glasslauncher.mods.networking.GlassNetworking.writeAndGetNbtLength(nbt, stream);
+        length = GlassNetworking.writeAndGetNbtLength(nbt, stream);
     }
 
     @Override
     public void apply(NetworkHandler networkHandler) {
-        BiConsumer<GlassPacket, NetworkHandler> packetHandler = net.glasslauncher.mods.networking.GlassNetworking.PACKET_HANDLERS.get(getFullId());
+        BiConsumer<GlassPacket, NetworkHandler> packetHandler = GlassNetworking.PACKET_HANDLERS.get(getFullId());
         if (packetHandler != null) {
             packetHandler.accept(this, networkHandler);
         }
