@@ -7,8 +7,9 @@ import net.glasslauncher.mods.gcapi3.impl.SeptFunction;
 import net.glasslauncher.mods.gcapi3.impl.object.ConfigEntryHandler;
 import net.glasslauncher.mods.gcapi3.impl.object.entry.EnumConfigEntryHandler;
 
-import java.lang.reflect.*;
-import java.util.function.*;
+import java.lang.reflect.Field;
+import java.lang.reflect.Type;
+import java.util.function.Function;
 
 public class ExampleConfigEnumFactories implements ConfigFactoryProvider {
 
